(#237)refactor: 질문 service interface 구현했습니다.
